# services/engine/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot Engine API
  description: |
    REST API for the Chatbot Engine — an NLP-powered chatbot platform that supports
    multi-group configuration, intent recognition, query execution, learning feedback,
    and administrative management.
  version: 1.0.0
  contact:
    name: Chatbot Engine Team
servers:
  - url: http://localhost:4000
    description: Local development
  - url: '{productionUrl}'
    description: Production
    variables:
      productionUrl:
        default: https://chatbot-engine.example.com
        description: Production server URL
...
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API key required for all `/api/admin/*` routes when the `ENGINE_API_KEY`
        environment variable is set. Can also be passed as a Bearer token in the
        Authorization header.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
...
    IntentItem:
      type: object
      properties:
        intent:
          type: string
          example: greeting
        utterances:
          type: array
          items:
            type: string
          example: [hello, hi there, hey]
        answers:
          type: array
          items:
            type: string
        utteranceCount:
          type: integer
    CreateIntentRequest:
      type: object
      required:
        - intent
        - utterances
      properties:
        intent:
          type: string
          example: greeting
        utterances:
          type: array
          items:
            type: string
        answers:
          type: array
          items:
            type: string

    # ── Admin: Entity ─────────────────────────────────────────────────────
...
paths:
...
  /api/admin/intents:
    get:
      tags: [Admin/Intents]
      summary: List all intents with utterances and entities
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Intents and entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  intents:
                    type: array
                    items:
                      $ref: '#/components/schemas/IntentItem'
                  entities:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        options:
                          type: object
                          additionalProperties:
                            type: array
                            items:
                              type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Intents]
      summary: Create or update an intent
      description: If the intent already exists, its utterances and answers are replaced.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '200':
          description: Intent saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  intent:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Intents]
      summary: Delete an intent
      security:
        - ApiKeyAuth: []
      parameters:
        - name: intent
          in: query
          required: true
          schema:
            type: string
          description: Intent name to delete
      responses:
        '200':
          description: Intent deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: string
        '400':
          description: Missing intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — ENTITIES
  # ════════════════════════════════════════════════════════════════════════
...